Feat - Add mutations of post, comment with error handling

+++ src/resolvers/Mutation.js
@@ -73,35 +73,138 @@ const Mutation = {
     )
   },
   createPost(parent, args, { prisma, request }, info) {
-    return prisma.mutation.createPost({
-      data: {
-        title: args.data.title,
-        body: args.data.body,
-        published: args.data.published,
+    const userId = getUserId(request)
+
+    if (!userId) {
+      throw new Error('User not found')
+    }
+
+    return prisma.mutation.createPost(
+      {
+        data: {
+          ...args.data,
+          author: {
+            connect: {
+              id: userId,
+            },
+          },
+        },
+      },
+      info,
+    )
+  },
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
       },
     })
+
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+
+    return prisma.mutation.deletePost(
+      {
+        where: {
+          id: args.id,
+        },
+      },
+      info,
+    )
   },
-  deletePost(parent, args, { prisma, request }, info) {
-    return prisma.mutation.deletePost({
-      where: {
-        id: args.id,
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
       },
     })
-  },
-  updatePost(parent, args, { prisma, request }, info) {
-    return prisma.mutation.updatePost({
-      where: {
-        id: args.id,
+
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
+
+    const allowedComments = await prisma.exists.Post({
+      id: args.id,
+      published: true,
+      allowComments: true,
+    })
+
+    if (
+      allowedComments &&
+      (args.data.published === false || args.data.allowComments === false)
+    ) {
+      await prisma.mutation.deleteManyComments({
+        where: {
+          post: {
+            id: args.id,
+          },
+        },
+      })
+    }
+
+    return prisma.mutation.updatePost(
+      {
+        where: {
+          id: args.id,
+        },
+        data: args.data,
       },
-      data: args.data,
-    })
+      info,
+    )
   },
-  createComment(parent, args, { prisma }, info) {
-    return prisma.mutation.createComment({
-      data: args.data,
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.data.post,
+      published: true,
+      allowComments: true,
     })
+
+    if (!postExists) {
+      throw new Error('Unable to find post')
+    }
+
+    return prisma.mutation.createComment(
+      {
+        data: {
+          text: args.data.text,
+          author: {
+            connect: {
+              id: userId,
+            },
+          },
+          post: {
+            connect: {
+              id: args.data.post,
+            },
+          },
+        },
+      },
+      info,
+    )
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to update comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id,
@@ -109,7 +212,20 @@ const Mutation = {
       data: args.data,
     })
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delete comment')
+    }
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id,
